Decompose MongoDB annotation bean-post-processors into per-annotation registrars
Split the 1314-line OccurrentBlockingAnnotationBeanPostProcessor and the 952-line OccurrentReactiveAnnotationBeanPostProcessor into thin coordinators plus package-private per-annotation registrars (subscription, projection, snapshot, and blocking-only saga) built over a StartPositionSupport helper. Behavior is unchanged: each coordinator keeps the single shared registeredIds registry, the exact afterSingletonsInstantiated ordering, and (blocking) the saga poller shutdown in destroy().

Lift the byte-identical shouldWaitUntilStarted(boolean, StartupMode) into the shared SubscriptionAnnotations. The genuinely divergent start-position logic (time-based starts, competing-consumer awareness, checkpoint API, Mongo template type) stays per stack.

Add container-free ApplicationContextRunner characterization tests for previously-untested error and validation branches (11 blocking, 9 reactor).
