+++ framework/spring-boot-starter-mongodb-reactive/src/main/java/org/occurrent/springboot/mongo/reactor/StartPositionSupport.java
@@ -1,0 +1,167 @@
+/*
+ *
+ *  Copyright 2026 Johan Haleby
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *         http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
+
+package org.occurrent.springboot.mongo.reactor;
+
+import org.occurrent.annotation.ResumeBehavior;
+import org.occurrent.annotation.StreamSubscription.StartPosition;
+import org.occurrent.eventstore.api.EventStoreCapability;
+import org.occurrent.eventstore.mongodb.spring.reactor.ReactorMongoEventStore;
+import org.occurrent.springboot.mongo.common.OccurrentProperties;
+import org.occurrent.springboot.mongo.common.SubscriptionAnnotations.StreamSubscriptionDefinition;
+import org.occurrent.subscription.DcbStartAt;
+import org.occurrent.subscription.GlobalCheckpoint;
+import org.occurrent.subscription.StartAt;
+import org.occurrent.subscription.api.reactor.CheckpointStorage;
+import org.occurrent.subscription.reactor.durable.ReactorDurableSubscriptionModel;
+import org.springframework.beans.factory.NoSuchBeanDefinitionException;
+import org.springframework.context.ApplicationContext;
+import org.springframework.data.mongodb.core.ReactiveMongoOperations;
+
+/**
+ * Start-position and startup logic for the reactive stack. The reactive catch-up model replays only by position (or,
+ * for DCB, by DCB position), so this diverges from the blocking stack's time-based start machinery: a
+ * {@link StreamSubscriptionDefinition} that asks for a specific historical time fails loud rather than resolving a
+ * wall-clock time to a position.
+ */
+class StartPositionSupport {
+
+    private final ApplicationContext applicationContext;
+
+    StartPositionSupport(ApplicationContext applicationContext) {
+        this.applicationContext = applicationContext;
+    }
+
+    void applyStartupWorkarounds() {
+        // These are workarounds for https://github.com/spring-projects/spring-framework/issues/32904
+        applicationContext.getBean(ReactiveMongoOperations.class);
+        try {
+            applicationContext.getBean("springApplicationAdminRegistrar");
+        } catch (NoSuchBeanDefinitionException ignored) {
+        }
+        // End workarounds
+    }
+
+    // A capability-agnostic subscription replays over the unified global position, so replay is supported whenever the
+    // store writes a position, regardless of which capabilities are enabled (unlike stream replay, which also requires
+    // the STREAM capability).
+    boolean positionReplaySupported() {
+        ReactorMongoEventStore eventStore = applicationContext.getBeanProvider(ReactorMongoEventStore.class).getIfAvailable();
+        return eventStore != null && eventStore.writesPosition();
+    }
+
+    // A @StreamSubscription can replay history when the store has the STREAM capability and writes stream position,
+    // which wires a catch-up model that replays stream filters by position. A combined STREAM+DCB store replays too. A
+    // DCB-only store also writes position but has no stream events, so it does not support stream history replay.
+    boolean streamHistoryReplaySupported() {
+        ReactorMongoEventStore eventStore = applicationContext.getBeanProvider(ReactorMongoEventStore.class).getIfAvailable();
+        if (eventStore == null || !eventStore.writesPosition()) {
+            return false;
+        }
+        OccurrentProperties occurrentProperties = applicationContext.getBean(OccurrentProperties.class);
+        return occurrentProperties.getEventStore().getCapabilities().contains(EventStoreCapability.STREAM);
+    }
+
+    // A stream subscription's start position. A specific start time (startAtISO8601 or startAtTimeEpochMillis) always
+    // fails loud, since position replay cannot resolve a wall-clock time to a position. BEGINNING_OF_TIME replays
+    // history when replay is supported (a STREAM store that writes position), and fails loud otherwise rather than
+    // silently starting live. NOW and DEFAULT are always supported.
+    StartAt generateStreamStartAt(StreamSubscriptionDefinition subscription, boolean historyReplaySupported) {
+        boolean specificTimeStart = !subscription.startAtISO8601().isBlank()
+                || subscription.startAtTimeEpochMillis() >= 0;
+        if (specificTimeStart) {
+            throw new IllegalArgumentException(("@StreamSubscription '%s' specifies a specific start time (startAtISO8601 or startAtTimeEpochMillis), but the reactive stack's position-based " +
+                    "stream catch-up cannot honor a specific historical start time, it can only replay from BEGINNING_OF_TIME, NOW, or DEFAULT. Use startAt = BEGINNING_OF_TIME to replay all history, " +
+                    "or NOW/DEFAULT, instead of a specific start time.").formatted(subscription.id()));
+        }
+        boolean beginningOfTimeStart = subscription.startAt() == StartPosition.BEGINNING_OF_TIME;
+        if (beginningOfTimeStart && !historyReplaySupported) {
+            throw new IllegalArgumentException(("@StreamSubscription '%s' asks to replay history (BEGINNING_OF_TIME), but this store does not support reactive stream history replay " +
+                    "(it has no STREAM capability, or stream position is off). Enable stream position (on by default) for a STREAM store, use startAt = NOW or DEFAULT, or use @DcbSubscription for a DCB store.").formatted(subscription.id()));
+        }
+        if (beginningOfTimeStart) {
+            // Map BEGINNING_OF_TIME to position 0, which the reactive stream catch-up model replays before going live.
+            return StartAt.checkpoint(GlobalCheckpoint.of(0));
+        }
+        return switch (subscription.startAt()) {
+            case NOW -> StartAt.now();
+            // DEFAULT resumes from the durably stored position, falling back to the subscription model default on first start.
+            case DEFAULT -> StartAt.subscriptionModelDefault();
+            case BEGINNING_OF_TIME -> throw new IllegalStateException("Unreachable: BEGINNING_OF_TIME handled above");
+        };
+    }
+
+    // Build the neutral StartAt over the unified global position. BEGINNING replays from global position 0,
+    // startAtGlobalPosition replays after a specific position, both applying the same replay-then-resume logic. NOW and
+    // DEFAULT go straight to live.
+    StartAt generateAgnosticStartAt(String subscriptionId, org.occurrent.annotation.StartPosition startPosition, long startAtGlobalPosition, ResumeBehavior resumeBehavior) {
+        if (startAtGlobalPosition >= 0) {
+            return replayThenResumeAgnostic(subscriptionId, StartAt.checkpoint(GlobalCheckpoint.of(startAtGlobalPosition)), resumeBehavior);
+        }
+        return switch (startPosition) {
+            case NOW -> StartAt.now();
+            case DEFAULT -> StartAt.subscriptionModelDefault();
+            case BEGINNING -> replayThenResumeAgnostic(subscriptionId, StartAt.checkpoint(GlobalCheckpoint.of(0)), resumeBehavior);
+        };
+    }
+
+    // Replay from replayStart, then on later restarts either resume from the stored position (DEFAULT) or replay again
+    // (SAME_AS_START_AT). SAME_AS_START_AT disables durable position storage by delegating to the parent subscription
+    // model, so an in-memory read model rebuilt on every boot sees every event and keeps no checkpoint. There is no
+    // reactive competing-consumer model, so only the durable layer is considered. Mirrors the DCB replayThenResume.
+    StartAt replayThenResumeAgnostic(String subscriptionId, StartAt replayStart, ResumeBehavior resumeBehavior) {
+        return switch (resumeBehavior) {
+            case SAME_AS_START_AT -> StartAt.dynamic(ctx -> {
+                boolean isDurableSubscription = ReactorDurableSubscriptionModel.class.isAssignableFrom(ctx.subscriptionModelType());
+                return isDurableSubscription ? null : replayStart;
+            });
+            case DEFAULT -> StartAt.dynamic(ctx -> {
+                CheckpointStorage storage = applicationContext.getBean(CheckpointStorage.class);
+                return storage.read(subscriptionId).blockOptional().isPresent() ? StartAt.subscriptionModelDefault() : replayStart;
+            });
+        };
+    }
+
+    DcbStartAt generateDcbStartAt(String subscriptionId, org.occurrent.annotation.StartPosition startPosition, long startAtDcbPosition, ResumeBehavior resumeBehavior) {
+        if (startAtDcbPosition >= 0) {
+            // Start after a specific position, applying the same replay-then-resume logic BEGINNING uses.
+            return replayThenResume(subscriptionId, DcbStartAt.afterPosition(startAtDcbPosition), resumeBehavior);
+        }
+        return switch (startPosition) {
+            case NOW -> DcbStartAt.now();
+            case DEFAULT -> DcbStartAt.subscriptionModelDefault();
+            case BEGINNING -> replayThenResume(subscriptionId, DcbStartAt.beginning(), resumeBehavior);
+        };
+    }
+
+    // Replay from replayStart, then on later restarts either resume from the stored position (DEFAULT) or replay again
+    // (SAME_AS_START_AT). SAME_AS_START_AT disables durable position storage by delegating to the parent subscription
+    // model, so an in-memory read model rebuilt on every boot sees every event and keeps no checkpoint. There is no
+    // reactive competing-consumer model, so only the durable layer is considered.
+    DcbStartAt replayThenResume(String subscriptionId, DcbStartAt replayStart, ResumeBehavior resumeBehavior) {
+        return switch (resumeBehavior) {
+            case SAME_AS_START_AT -> DcbStartAt.dynamic(ctx -> {
+                boolean isDurableSubscription = ReactorDurableSubscriptionModel.class.isAssignableFrom(ctx.subscriptionModelType());
+                return isDurableSubscription ? null : replayStart;
+            });
+            case DEFAULT -> DcbStartAt.dynamic(ctx -> {
+                CheckpointStorage storage = applicationContext.getBean(CheckpointStorage.class);
+                return storage.read(subscriptionId).blockOptional().isPresent() ? DcbStartAt.subscriptionModelDefault() : replayStart;
+            });
+        };
+    }
+}
